try as python shell
